+++ src/node-lib/curriculum-api-2023/queries/curriculumOverview/curriculumOverview.query.ts
@@ -1,22 +1,21 @@
-// import curriculumOverviewSchema from "./curriculumOverview.schema";
+import curriculumOverviewSchema from "./curriculumOverview.schema";
 
 import OakError from "@/errors/OakError";
 import { Sdk } from "@/node-lib/curriculum-api-2023/sdk";
-import { curriculumOverviewMVFixture } from "@/node-lib/curriculum-api-2023/fixtures/curriculumOverview.fixture";
 
 const curriculumOverviewQuery =
-  (sdk: Sdk) =>
-  async (args: {
-    phaseSlug: string;
-    subjectSlug: string;
-    examboardSlug: string | null;
-  }) => {
-    console.log(sdk, args);
+  (sdk: Sdk) => async (args: { phaseSlug: string; subjectSlug: string }) => {
     if (!args.phaseSlug || !args.subjectSlug) {
       throw new OakError({ code: "curriculum-api/not-found" });
     }
+    const res = await sdk.curriculumOverview(args);
+    const [curriculumOverview] = res.curriculumOverview;
 
-    return curriculumOverviewMVFixture();
+    if (!curriculumOverview) {
+      throw new OakError({ code: "curriculum-api/not-found" });
+    }
+
+    return curriculumOverviewSchema.parse(curriculumOverview);
   };
 
 export default curriculumOverviewQuery;